@@ -3,9 +3,9 @@
  */
 package com.pc.android.sdk.internal;
 
+import android.app.AlertDialog;
 import android.content.Context;
 import android.content.DialogInterface;
-import android.support.v7.app.AlertDialog;
 import android.view.View;
 import android.webkit.CookieManager;
 import android.webkit.WebView;
